:art: Extract columns from UI
