+++ lib/ui-atom-adapter.js
@@ -1,8 +1,12 @@
 'use babel'
 
+import {React} from 'react-for-atom'
 import Bacon from 'baconjs'
 import * as atoms from './atom-streams'
 import DisposableValues from './disposable-values'
+import DateTimeComponent from './react/cells/date-time-component'
+import SummaryComponent from './react/cells/summary-component'
+import TagsComponent from './react/cells/tags-component'
 import UI from './ui'
 
 // Adapter for Atom-land
@@ -34,9 +38,48 @@ export default class UIAtomAdapter {
       .debounce(50)
       .merge(project.isLoadingFilesProp.changes()) // placed after debounce, to trigger immeditately
 
+    const columns = []
+    columns.push({
+      title: 'Name',
+      width: 50,
+      createCell: (file, state) => {
+        return <SummaryComponent key='name' file={file}
+        searchStr={state.searchStr} tokens={state.results.tokens} />
+      }
+    })
+    if (project.darwin) {
+      columns.push({
+        title: 'Tags',
+        width: 20,
+        createCell: (file, state, isSelected) => {
+          const saveTags = tags => project.darwin.setTags(file.path, tags.trim().split(' '))
+          return (
+            <TagsComponent key='tags' tags={file.tags} isSelected={isSelected} saveTags={saveTags} />
+          )
+        }
+      })
+    }
+    columns.push({
+      title: 'Date modified',
+      width: 15,
+      createCell: file => {
+        return <DateTimeComponent key='mtime' time={file.stat.mtime} />
+      }
+    })
+    columns.push({
+      title: 'Date created',
+      width: 15,
+      createCell: file => {
+        return <DateTimeComponent key='birhtime' time={file.stat.birthtime} />
+      }
+    })
+
     const ui = new UI({
       domNode: panelDomNode,
-      project: project,
+      isLoadingFilesProp: project.isLoadingFilesProp,
+      filesProp: project.filesProp,
+      resultsProp: project.resultsProp,
+      columns: columns,
       panelHeightStream: atoms.createConfigStream('textual-velocity.panelHeight'),
       focusOnSearchStream: focusOnSearchStream,
       deselectStream: Bacon.mergeAll(
