+++ lib/ui.js
@@ -4,9 +4,6 @@ import {React, ReactDOM} from 'react-for-atom'
 import Bacon from 'baconjs'
 import R from 'ramda'
 import PanelComponent from './react/panel-component'
-import SummaryComponent from './react/cells/summary-component'
-import DateTimeComponent from './react/cells/date-time-component'
-import TagsComponent from './react/cells/tags-component'
 
 // UI class, representing the presentation logic
 //
@@ -14,7 +11,7 @@ import TagsComponent from './react/cells/tags-component'
 // Basically it should be able to work in any context, least the provided streams fulfills the contract.
 export default class UI {
 
-  constructor ({domNode, project, panelHeightStream, focusOnSearchStream, deselectStream}) {
+  constructor ({domNode, isLoadingFilesProp, filesProp, resultsProp, columns, panelHeightStream, focusOnSearchStream, deselectStream}) {
     this._domNode = domNode
 
     const keyDownBus = new Bacon.Bus()
@@ -33,7 +30,7 @@ export default class UI {
         [escKeyStream], R.always({}),
 
         // Selected next item
-        [downKeyStream, project.filesProp, project.resultsProp], (current, _, files, {items}) => {
+        [downKeyStream, filesProp, resultsProp], (current, _, files, {items}) => {
           const i = R.defaultTo(-1, current.index) + 1 // start on first item
           const nextItem = items[i]
           if (nextItem) {
@@ -47,7 +44,7 @@ export default class UI {
           }
         },
         // Select prev item
-        [upKeyStream, project.filesProp, project.resultsProp], (current, _, files, {items}) => {
+        [upKeyStream, filesProp, resultsProp], (current, _, files, {items}) => {
           const i = R.defaultTo(items.length, current.index) - 1 // start on last item
           const prevItem = items[i]
           if (prevItem) {
@@ -61,7 +58,7 @@ export default class UI {
           }
         },
         // User clicks on one of the visible item, change to that one directly
-        [selectIndexBus, project.filesProp, project.resultsProp], (_, i, files, {items}) => {
+        [selectIndexBus, filesProp, resultsProp], (_, i, files, {items}) => {
           const item = items[i]
           return {
             item: item,
@@ -70,7 +67,7 @@ export default class UI {
           }
         },
         // Files changes (add/delete/order), update index from known path from current
-        [project.filesProp.changes(), project.resultsProp], (current, files, {items}) => {
+        [filesProp.changes(), resultsProp], (current, files, {items}) => {
           const currentPath = current.path
           if (currentPath) {
             const i = items.findIndex(({id}) => files[id].path === currentPath)
@@ -85,48 +82,12 @@ export default class UI {
         }
       ).skipDuplicates()
 
-    const columns = []
-    columns.push({
-      title: 'Name',
-      width: 50,
-      createCell: (file, state) => {
-        return <SummaryComponent key='name' file={file}
-        searchStr={state.searchStr} tokens={state.results.tokens} />
-      }
-    })
-    if (project.darwin) {
-      columns.push({
-        title: 'Tags',
-        width: 20,
-        createCell: (file, state, isSelected) => {
-          const saveTags = tags => project.darwin.setTags(file.path, tags.trim().split(' '))
-          return (
-            <TagsComponent key='tags' tags={file.tags} isSelected={isSelected} saveTags={saveTags} />
-          )
-        }
-      })
-    }
-    columns.push({
-      title: 'Date modified',
-      width: 15,
-      createCell: file => {
-        return <DateTimeComponent key='mtime' time={file.stat.mtime} />
-      }
-    })
-    columns.push({
-      title: 'Date created',
-      width: 15,
-      createCell: file => {
-        return <DateTimeComponent key='birhtime' time={file.stat.birthtime} />
-      }
-    })
-
     const reactPanel = ReactDOM.render(
       <PanelComponent focusOnSearchStream={focusOnSearchStream.merge(selectIndexBus.map(true))}
         panelHeightStream={panelHeightStream}
-        filesProp={project.filesProp}
-        resultsProp={project.resultsProp}
-        isLoadingFilesProp={project.isLoadingFilesProp}
+        filesProp={filesProp}
+        resultsProp={resultsProp}
+        isLoadingFilesProp={isLoadingFilesProp}
         selectIndexBus={selectIndexBus}
         selectedIndexStream={selectedProp.changes().map('.index')}
         keyDownBus={keyDownBus}
@@ -136,7 +97,7 @@ export default class UI {
     this.panelHeightProp = reactPanel.panelHeightProp
     this.searchProp = reactPanel.searchProp
     this.selectedFileProp = selectedProp
-      .combine(project.filesProp, (selected, files) => {
+      .combine(filesProp, (selected, files) => {
         const id = R.path(['item', 'id'], selected)
         return files[id]
       })
